+++ src/internal/copy-conditions.ts
